backend_api/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
flwr==1.5.0
torch==2.0.1
numpy==1.24.3
pandas==2.0.3
cryptography==41.0.3
